// projects/venus/src/lib/types/tokens.ts
export interface VenusTokens {
  colors?: {
    background?: string;
    primaryLow?: string;
    primaryMid?: string;
    primaryDark?: string;
    secondaryLow?: string;
    secondaryLight?: string;
    secondaryRegular?: string;
    secondaryMid?: string;
    secondaryDark?: string;
    success?: string;
    successHover?: string;
    danger?: string;
    dangerHover?: string;
    warning?: string;
    warningHover?: string;
    blue?: string;
    blueHover?: string;
    purple?: string;
    purpleHover?: string;
    black?: string;
    white?: string;
    gray?: string;
  };

  fontSizes?: {
    xs?: string;
    sm?: string;
    md?: string;
    lg?: string;
    xl?: string;
    xxl?: string;
    xxxl?: string;
  };

  fontWeights?: {
    regular?: string;
    medium?: string;
    bold?: string;
  };

  radii?: {
    sm?: string;
    md?: string;
    lg?: string;
    xl?: string;
  };
}